+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+next [????.??.??]
+-----------------
+* Replace `test-framework` with `tasty` in the test suite.
+
 0.3.2 [2024.12.04]
 ------------------
 * Drop support for pre-8.0 versions of GHC.